Create separate wiki page base class with Wikitext support.
This should help in making custom page types.
