+++ src/fitnesse/wiki/SymbolicPage.java
@@ -10,8 +10,7 @@ import java.util.List;
 import fitnesse.wikitext.parser.ParsingPage;
 import fitnesse.wikitext.parser.Symbol;
 
-public class SymbolicPage extends BaseWikiPage {
-  private static final long serialVersionUID = 1L;
+public class SymbolicPage extends BaseWikitextPage {
 
   public static final String PROPERTY_NAME = "SymbolicLinks";
 
@@ -25,6 +24,10 @@ public class SymbolicPage extends BaseWikiPage {
 
   public WikiPage getRealPage() {
     return realPage;
+  }
+
+  public boolean containsWikitext() {
+    return realPage instanceof WikitextPage;
   }
 
   @Override
@@ -86,7 +89,7 @@ public class SymbolicPage extends BaseWikiPage {
 
   @Override
   public String getVariable(String name) {
-    if (realPage instanceof WikitextPage) {
+    if (containsWikitext()) {
       return super.getVariable(name);
     }
     return realPage.getVariable(name);
@@ -94,7 +97,7 @@ public class SymbolicPage extends BaseWikiPage {
 
   @Override
   public String getHtml() {
-    if (realPage instanceof WikitextPage) {
+    if (containsWikitext()) {
       return super.getHtml();
     }
     return realPage.getHtml();
@@ -102,7 +105,7 @@ public class SymbolicPage extends BaseWikiPage {
 
   @Override
   public ParsingPage getParsingPage() {
-    if (realPage instanceof WikitextPage) {
+    if (containsWikitext()) {
       return super.getParsingPage();
     }
     return null;
@@ -110,7 +113,7 @@ public class SymbolicPage extends BaseWikiPage {
 
   @Override
   public Symbol getSyntaxTree() {
-    if (realPage instanceof WikitextPage) {
+    if (containsWikitext()) {
       return super.getSyntaxTree();
     }
     return Symbol.emptySymbol;
