+++ src/fitnesse/wiki/WikitextPage.java
@@ -4,7 +4,7 @@ import fitnesse.wikitext.parser.ParsingPage;
 import fitnesse.wikitext.parser.Symbol;
 
 /**
- * This interface denotes a class that can expose parsed wiki page content,
+ * This interface denotes a class that can expose parsed wiki page content.
  */
 public interface WikitextPage {
   Symbol getSyntaxTree();
