+++ src/fitnesse/wiki/fs/ExternalSuitePage.java
@@ -1,5 +1,6 @@
 package fitnesse.wiki.fs;
 
+import fitnesse.wiki.BaseWikitextPage;
 import fitnesse.wiki.WikiPageProperties;
 import fitnesse.wikitext.parser.VariableSource;
 
@@ -9,20 +10,19 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-import fitnesse.wiki.BaseWikiPage;
 import fitnesse.wiki.PageData;
 import fitnesse.wiki.PageType;
 import fitnesse.wiki.VersionInfo;
 import fitnesse.wiki.WikiPage;
 import fitnesse.util.Clock;
 
-public class ExternalSuitePage extends BaseWikiPage {
+public class ExternalSuitePage extends BaseWikitextPage {
   public static final String HTML = ".html";
 
   private File path;
   private FileSystem fileSystem;
 
-  public ExternalSuitePage(File path, String name, BaseWikiPage parent, FileSystem fileSystem, VariableSource variableSource) {
+  public ExternalSuitePage(File path, String name, WikiPage parent, FileSystem fileSystem, VariableSource variableSource) {
     super(name, parent, variableSource);
     this.path = path;
     this.fileSystem = fileSystem;
