+++ src/fitnesse/wiki/fs/ExternalSuitePageFactory.java
@@ -2,12 +2,12 @@ package fitnesse.wiki.fs;
 
 import java.io.File;
 
-import fitnesse.wiki.BaseWikiPage;
+import fitnesse.wiki.BaseWikitextPage;
 import fitnesse.wiki.WikiPage;
 import fitnesse.wiki.WikiPageFactory;
 import fitnesse.wikitext.parser.VariableSource;
 
-public class ExternalSuitePageFactory implements WikiPageFactory<BaseWikiPage> {
+public class ExternalSuitePageFactory implements WikiPageFactory<WikiPage> {
 
   private final FileSystem fileSystem;
 
@@ -16,7 +16,7 @@ public class ExternalSuitePageFactory implements WikiPageFactory<BaseWikiPage> {
   }
 
   @Override
-  public WikiPage makePage(File path, String pageName, BaseWikiPage parent, VariableSource variableSource) {
+  public WikiPage makePage(File path, String pageName, WikiPage parent, VariableSource variableSource) {
     return new ExternalSuitePage(path, pageName, parent, fileSystem, variableSource);
   }
