+++ src/fitnesse/wiki/fs/ExternalTestPage.java
@@ -6,20 +6,20 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-import fitnesse.wiki.BaseWikiPage;
+import fitnesse.util.Clock;
+import fitnesse.wiki.BaseWikitextPage;
 import fitnesse.wiki.PageData;
 import fitnesse.wiki.PageType;
 import fitnesse.wiki.VersionInfo;
 import fitnesse.wiki.WikiPage;
 import fitnesse.wiki.WikiPageProperties;
 import fitnesse.wikitext.parser.VariableSource;
-import fitnesse.util.Clock;
 
-public class ExternalTestPage extends BaseWikiPage {
+public class ExternalTestPage extends BaseWikitextPage {
   private FileSystem fileSystem;
   private File path;
 
-  public ExternalTestPage(File path, String name, BaseWikiPage parent, FileSystem fileSystem, VariableSource variableSource) {
+  public ExternalTestPage(File path, String name, WikiPage parent, FileSystem fileSystem, VariableSource variableSource) {
     super(name, parent, variableSource);
     this.path = path;
     this.fileSystem = fileSystem;
